requirements.txt
streamlit==1.22.0
tensorflow==2.13.0
numpy
pandas
scikit-learn
matplotlib
Pillow
opencv-python
plotly
